server-py/requirements.txt
flask>=1.1.2
flask-cors>=3.0.9
werkzeug>=1.0.1
gevent>=20.9.0
pandas>=1.1.1
xlrd>=2.0.0
openpyxl>=3.0.6
pyarrow>=2.0.0
sqlalchemy>=1.3.17
psycopg2-binary>=2.8.5
mysqlclient>=2.0.3
pymssql>=2.1.5
python-dotenv
